feat(EFG-60): Trocando nome de grupo do projeto.
